+++ buildDocs.ps1
@@ -1,0 +1,2 @@
+dotnet tool update -g docfx
+docfx .\docs\docfx.json --serve