+++ src/idunno.Bluesky/Post.cs
@@ -75,19 +75,19 @@ namespace idunno.Bluesky
             }
         }
 
-
         /// <summary>
         /// Creates a new instance of <see cref="Post"/>.
         /// </summary>
         /// <param name="text">The text for the post.</param>
         /// <param name="lang">The language the post is written in.</param>
-        /// <exception cref="ArgumentNullException">Thrown when <paramref name="text" /> or <paramref name="lang" /> is null or empty.</exception>
+        /// <exception cref="ArgumentNullException">Thrown when <paramref name="text" /> is null or empty.</exception>
+        /// <exception cref="ArgumentException">Thrown when <paramref name="lang" /> is empty.</exception>
         public Post(
             string text,
             string lang) : this(text: text, langs: new List<string>() { lang })
         {
-            ArgumentNullException.ThrowIfNullOrEmpty(text);
-            ArgumentNullException.ThrowIfNullOrEmpty(lang);
+            ArgumentException.ThrowIfNullOrEmpty(text);
+            ArgumentException.ThrowIfNullOrEmpty(lang);
         }
 
         /// <summary>
@@ -95,13 +95,13 @@ namespace idunno.Bluesky
         /// </summary>
         /// <param name="text">The text for the post.</param>
         /// <param name="langs">The languages the post is written in.</param>
-        /// <exception cref="ArgumentNullException">Thrown when <paramref name="text" /> is null or empty, or <paramref name="langs"/> is null.</exception>
+        /// <exception cref="ArgumentNullException">Thrown when <paramref name="text" /> is null or empty.</exception>
         /// <exception cref="ArgumentOutOfRangeException">Thrown when <paramref name="langs"/> is an empty list.</exception>
         public Post(
             string text,
             IList<string> langs) : this(text: text, facets: null, langs: langs, embed: null, reply: null, labels : null, tags : null)
         {
-            ArgumentNullException.ThrowIfNullOrEmpty(text);
+            ArgumentException.ThrowIfNullOrEmpty(text);
             ArgumentNullException.ThrowIfNull(langs);
 
             ArgumentOutOfRangeException.ThrowIfZero(langs.Count);
@@ -182,6 +182,7 @@ namespace idunno.Bluesky
             Reply = reply;
             Facets = facets;
             Langs = langs;
+
             Embed = embed;
 
             if (labels is not null)
@@ -213,6 +214,142 @@ namespace idunno.Bluesky
 
                 Tags = new List<string>(tags).AsReadOnly();
             }
+        }
+
+        /// <summary>
+        /// Creates a new instance of <see cref="Post"/>.
+        /// </summary>
+        /// <param name="text">The text for the post.</param>
+        /// <param name="reply">The <see cref="ReplyReferences"/>, if any, of the post this post is in reply to.</param>
+        /// <param name="facets">A collection of <see cref="Facet"/>s for the post.</param>
+        /// <param name="langs">A collection of language strings, if any, that the post is written in.</param>
+        /// <param name="image">The <see cref="EmbeddedImage"/> to embed in the post..</param>
+        /// <param name="labels">A collection of <see cref="SelfLabels"/> to apply to the post, if any.</param>
+        /// <param name="tags">A collection of tags to apply to the post, if any.</param>
+        /// <exception cref="ArgumentNullException">Thrown when <paramref name="image"/> is null.</exception>
+        public Post(
+            string? text,
+            EmbeddedImage image,
+            IList<Facet>? facets = null,
+            IList<string>? langs = null,
+            ReplyReferences? reply = null,
+            SelfLabels? labels = null,
+            IReadOnlyCollection<string>? tags = null)
+            : this(text,
+                createdAt: DateTimeOffset.UtcNow,
+                facets: facets,
+                langs: langs,
+                new EmbeddedImages(new List<EmbeddedImage>() { image }),
+                reply: reply,
+                labels: labels,
+                tags: tags)
+        {
+            ArgumentNullException.ThrowIfNull(image);
+        }
+
+        /// <summary>
+        /// Creates a new instance of <see cref="Post"/>.
+        /// </summary>
+        /// <param name="text">The text for the post.</param>
+        /// <param name="reply">The <see cref="ReplyReferences"/>, if any, of the post this post is in reply to.</param>
+        /// <param name="facets">A collection of <see cref="Facet"/>s for the post.</param>
+        /// <param name="langs">A collection of language strings, if any, that the post is written in.</param>
+        /// <param name="image">The <see cref="EmbeddedImage"/> to embed in the post..</param>
+        /// <param name="labels">A collection of <see cref="SelfLabels"/> to apply to the post, if any.</param>
+        /// <param name="tags">A collection of tags to apply to the post, if any.</param>
+        /// <param name="createdAt">The <see cref="DateTimeOffset"/> the post was created on.</param>
+        /// <exception cref="ArgumentNullException">Thrown when <paramref name="image"/> is null.</exception>
+        public Post(
+            string? text,
+            DateTimeOffset createdAt,
+            EmbeddedImage image,
+            IList<Facet>? facets = null,
+            IList<string>? langs = null,
+            ReplyReferences? reply = null,
+            SelfLabels? labels = null,
+            IReadOnlyCollection<string>? tags = null)
+            : this (text,
+                createdAt: createdAt,
+                facets: facets,
+                langs: langs,
+                new EmbeddedImages(new List<EmbeddedImage>() { image }),
+                reply : reply,
+                labels: labels,
+                tags: tags)
+        {
+            ArgumentNullException.ThrowIfNull(image);
+        }
+
+        /// <summary>
+        /// Creates a new instance of <see cref="Post"/>.
+        /// </summary>
+        /// <param name="text">The text for the post.</param>
+        /// <param name="reply">The <see cref="ReplyReferences"/>, if any, of the post this post is in reply to.</param>
+        /// <param name="facets">A collection of <see cref="Facet"/>s for the post.</param>
+        /// <param name="langs">A collection of language strings, if any, that the post is written in.</param>
+        /// <param name="images">The <see cref="EmbeddedImage"/> to embed in the post..</param>
+        /// <param name="labels">A collection of <see cref="SelfLabels"/> to apply to the post, if any.</param>
+        /// <param name="tags">A collection of tags to apply to the post, if any.</param>
+        /// <exception cref="ArgumentNullException">Thrown when <paramref name="images"/> is null.</exception>
+        /// <exception cref="ArgumentOutOfRangeException">Thrown when <paramref name="images"/> or empty, or contains more than the maximum allowed number of images.</exception>
+        public Post(
+            string? text,
+            ICollection<EmbeddedImage> images,
+            IList<Facet>? facets = null,
+            IList<string>? langs = null,
+            ReplyReferences? reply = null,
+            SelfLabels? labels = null,
+            IReadOnlyCollection<string>? tags = null)
+            : this(text,
+                createdAt: DateTimeOffset.UtcNow,
+                facets: facets,
+                langs: langs,
+                new EmbeddedImages(images),
+                reply: reply,
+                labels: labels,
+                tags: tags)
+        {
+            ArgumentNullException.ThrowIfNull(images);
+
+            ArgumentOutOfRangeException.ThrowIfZero(images.Count);
+            ArgumentOutOfRangeException.ThrowIfGreaterThan(images.Count, Maximum.ImagesInPost);
+        }
+
+        /// <summary>
+        /// Creates a new instance of <see cref="Post"/>.
+        /// </summary>
+        /// <param name="text">The text for the post.</param>
+        /// <param name="reply">The <see cref="ReplyReferences"/>, if any, of the post this post is in reply to.</param>
+        /// <param name="facets">A collection of <see cref="Facet"/>s for the post.</param>
+        /// <param name="langs">A collection of language strings, if any, that the post is written in.</param>
+        /// <param name="images">The <see cref="EmbeddedImage"/> to embed in the post..</param>
+        /// <param name="labels">A collection of <see cref="SelfLabels"/> to apply to the post, if any.</param>
+        /// <param name="tags">A collection of tags to apply to the post, if any.</param>
+        /// <param name="createdAt">The <see cref="DateTimeOffset"/> the post was created on.</param>
+        /// <exception cref="ArgumentNullException">Thrown when <paramref name="images"/> is null.</exception>
+        /// <exception cref="ArgumentOutOfRangeException">Thrown when <paramref name="images"/> or empty, or contains more than the maximum allowed number of images.</exception>
+        public Post(
+            string? text,
+            DateTimeOffset createdAt,
+            ICollection<EmbeddedImage> images,
+            IList<Facet>? facets = null,
+            IList<string>? langs = null,
+            ReplyReferences? reply = null,
+            SelfLabels? labels = null,
+            IReadOnlyCollection<string>? tags = null)
+            : this(text,
+                createdAt: createdAt,
+                facets: facets,
+                langs: langs,
+                new EmbeddedImages(images),
+                reply: reply,
+                labels: labels,
+                tags: tags)
+        {
+            ArgumentNullException.ThrowIfNull(images);
+
+            ArgumentOutOfRangeException.ThrowIfZero(images.Count);
+            ArgumentOutOfRangeException.ThrowIfGreaterThan(images.Count, Maximum.ImagesInPost);
         }
 
         /// <summary>