+++ history.md
@@ -1,28 +1,59 @@
 # idunno.Bluesky Version History
 
-## 0.2.0
+## 0.1.3
 
-### idunno.Bluesky
+### Features
 
-* Adds
-  * Profile editing
+#### idunno.Bluesky
 
-## 0.1.1 - Bug Fixes
+* Add self labels for posts
+* *Breaking* - Consolidation of record value classes
 
-### idunno.AtProto
+### Bug fixes
+
+#### idunno.Bluesky
+
+* Fixed facet positioning
+
+### Docs
+
+* Added Profile editing sample
+
+## 0.1.2
+
+### Features
+
+#### idunno.Bluesky
+
+* Adds Profile editing
+
+### Bug fixes
+
+#### idunno.Bluesky
+
+* Fixed positioning bug for PostBuilder facets
+* Removed incorrect link length check.
+
+## 0.1.1
+
+### Bug Fixes
+
+#### idunno.AtProto
 
 * Fixed bug when making requests a personal PDS timed out after authentication.
 
-## 0.1.0 - Initial Release
+## 0.1.0
 
-### idunno.AtProto
+### Features
+
+#### idunno.AtProto
 
 * PDS authentication and session management
 * List, Create, Get, Put, Delete records
 * Blob uploads
 * Handle and PDS resolution
 
-### idunno.Bluesky
+#### idunno.Bluesky
 
 * Viewing feeds
 * Viewing a user's timeline and notifications